requirements.txt
google-cloud-bigquery
google-auth
google-cloud-storage

